+++ src/content/flora/GrassFactory.cpp
@@ -45,7 +45,7 @@ FloraProperties GrassFactory::getProperties(uint16_t speciesId) const {
     .blockId = def.blockId,
     .name = def.name,
     .renderType = def.isTall ? FloraRenderType::TALL_CROSS_QUAD : FloraRenderType::CROSS_QUAD,
-    .textureLayers = {static_cast<uint8_t>(def.textureLayer)},
+    .textureLayers = {static_cast<uint16_t>(def.textureLayer)},
     .acceptableSoil = def.acceptableSoil,
     .biomeAffinity = def.biomeAffinity,
     .lightRequirements = {8, 0, 15},
